requirements.txt
gradio==4.25.0
gtts==2.5.1
playsound==1.3.0
python-dotenv==1.0.1
sounddevice==0.4.6
scipy==1.13.1
numpy==1.26.4
groq==0.5.0
elevenlabs==2.8.1
